Split planner function return body lowering
- Move tail-aware return body conversion into planner::function::return_body.
- Keep planner::function focused on function body planning and assembly.
- Colocate return-family mismatch tests with the moved return body logic.
